requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymupdf==1.22.3
numpy==1.24.4